# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bastion",
    "description": "ERP para pyme española. Contrato de la versión 1 de la API, bajo /api/v1/{modulo}/{recurso}. Se genera al compilar desde los controladores y los DTO: no se escribe a mano, y el cliente del frontal se genera de él.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/organizacion/empresas/{id}": {
      "get": {
        "tags": [
          "Empresas"
        ],
        "summary": "Devuelve una empresa.",
        "operationId": "Empresas_Obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la empresa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "put": {
        "tags": [
          "Empresas"
        ],
        "summary": "Cambia los datos de una empresa.",
        "operationId": "Empresas_Modificar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la empresa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarEmpresaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarEmpresaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModificarEmpresaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      },
      "delete": {
        "tags": [
          "Empresas"
        ],
        "summary": "Bloquea una empresa (R16). No la borra.",
        "description": "El verbo es `DELETE` porque es lo que el cliente quiere decir —«quítame esto de en\nmedio»— y lo que se hace por debajo es bloquear: una empresa puede ser un empresario\nindividual, y el art. 32 de la LOPDGDD manda bloquear, no destruir.",
        "operationId": "Empresas_Bloquear",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la empresa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/v1/organizacion/empresas/{id}/desbloqueo": {
      "post": {
        "tags": [
          "Empresas"
        ],
        "summary": "Devuelve una empresa bloqueada a la actividad.",
        "description": "    No existe el DELETE inverso, así que es un POST a un sub-recurso. El\n      estado de la ficha no es un recurso que se sustituya con PUT: es el desenlace de una\n      operación con nombre propio, y por eso tiene permiso propio.\n    En el 0.4 esta operación existía en el dominio y no tenía puerta HTTP, porque abrirla sin\nautorización habría dejado a cualquiera revirtiendo bloqueos del art. 32. Con el permiso\ndetrás, ese motivo desaparece.",
        "operationId": "Empresas_Desbloquear",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador de la empresa.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DireccionDto": {
        "required": [
          "calle",
          "codigoPostal",
          "poblacion",
          "pais"
        ],
        "type": "object",
        "properties": {
          "calle": {
            "maxLength": 70,
            "minLength": 0,
            "type": "string",
            "description": "Nombre de la vía."
          },
          "numero": {
            "maxLength": 16,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Número, portal, piso. Opcional: hay direcciones que no lo tienen."
          },
          "codigoPostal": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string",
            "description": "Código postal, tal como lo escribe el país de la dirección."
          },
          "poblacion": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "description": "Población."
          },
          "subdivision": {
            "maxLength": 35,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ],
            "description": "Provincia, región o estado. Opcional: no todos los países la usan."
          },
          "pais": {
            "maxLength": 2,
            "minLength": 2,
            "type": "string",
            "description": "País en ISO 3166-1 alfa-2, dos letras."
          }
        },
        "description": "Una dirección estructurada, en los seis campos que exige R17."
      },
...
      "EmpresaDto": {
        "required": [
          "id",
          "nif",
          "razonSocial",
          "domicilioFiscal",
          "divisaBase",
          "regimenDeIva"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador de la empresa.",
            "format": "uuid"
          },
          "nif": {
            "type": "string",
            "description": "NIF, ya normalizado."
          },
          "razonSocial": {
            "type": "string",
            "description": "Razón social o nombre del empresario individual."
          },
          "domicilioFiscal": {
            "description": "Domicilio fiscal, estructurado (R17).",
            "$ref": "#/components/schemas/DireccionDto"
          },
          "divisaBase": {
            "type": "string",
            "description": "Divisa base en ISO 4217."
          },
          "regimenDeIva": {
            "type": "string",
            "description": "Régimen de IVA, como texto."
          }
        },
        "description": "Una empresa, tal como sale de la API."
      },
...
      "ModificarEmpresaDto": {
        "required": [
          "razonSocial",
          "domicilioFiscal",
          "divisaBase",
          "regimenDeIva"
        ],
        "type": "object",
        "properties": {
          "razonSocial": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "description": "Razón social, o nombre del empresario individual."
          },
          "domicilioFiscal": {
            "description": "Domicilio fiscal, en los seis campos de R17.",
            "$ref": "#/components/schemas/DireccionDto"
          },
          "divisaBase": {
            "maxLength": 3,
            "minLength": 3,
            "type": "string",
            "description": "Divisa base en ISO 4217, tres letras."
          },
          "regimenDeIva": {
            "type": "string",
            "description": "Régimen de IVA, como texto."
          }
        },
        "description": "Lo que se puede cambiar de una empresa ya dada de alta."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "testigoDeAcceso": {
        "type": "http",
        "description": "Testigo de acceso de quince minutos, en la cabecera Authorization. Lo emite POST /api/v1/identidad/sesiones y lo renueva POST .../sesiones/renovacion, que no lo lee de aquí sino de una cookie HttpOnly. El testigo de refresco no aparece en este contrato a propósito: el navegador no tiene que poder leerlo.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}